backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
cloudinary>=1.32
django-cloudinary-storage>=1.6
django-filter>=23.2
dj-rest-auth>=2.2
dj-rest-auth[with_social]>=2.2
django-allauth>=0.52.0
djangorestframework-simplejwt>=5.2
